Working with home assistent
